Final Push - Ready for Review
Added the table cells that include the calendar days. Added functions to make sure that the day of the week starts and ends correctly, as well as display empty cells around those. Added in the events for each cell, by including another script element. Added that today's cell will be highlighted on the calendar. Added extra code to determine whether a year is truly a leap year, with extra modulus operators.